added a function to us symbols and removed a background task
