+++ backend/money_maker/tasks/task.py
@@ -57,25 +57,3 @@ def update_yh_stocks():
 
     db.session.execute(on_conflict_statement, formatted_yh_information)
     db.session.commit()
-
-
-# noinspection PyTypeChecker
-@shared_task
-def get_american_yh_stocks():
-    stock_data = PyTickerSymbols()
-
-    sp500_yahoo = stock_data.get_sp_500_nyc_yahoo_tickers()
-    nasdaq_yahoo = stock_data.get_nasdaq_100_nyc_yahoo_tickers()
-    dow_jones = stock_data.get_dow_jones_nyc_yahoo_tickers()
-
-    result = sp500_yahoo + nasdaq_yahoo + dow_jones
-    result_set = ([{"symbol": element} for element in (set(result))])
-
-    stmt = insert(tP).values(result_set)
-
-    on_conflict_ignore = stmt.on_conflict_do_nothing(
-        index_elements=['symbol']
-    )
-
-    db.session.execute(on_conflict_ignore)
-    db.session.commit()
